Refactor parser into submodules and move tests
